Fix bugs in queries

+++ MigrationsService/Migrations/script-0022 create_procedure-get_group_total.sql
@@ -3,7 +3,7 @@ CREATE FUNCTION GetGroupTotal(in_group_id BIGINT)
 RETURNS INT AS $$
 BEGIN
 -- One group can only have one larpake
-RETURN (SELECT SUM(e.points)
+RETURN (SELECT SUM(e.points)::INT
         FROM freshman_groups g
             LEFT JOIN freshman_group_members m
                 ON g.id = m.group_id
